requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
apscheduler>=3.10.0
pyTelegramBotAPI>=4.18.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
gspread>=6.0.0
google-auth>=2.28.0
